+++ cyral/internal/policy/set/constants.go
@@ -1,0 +1,6 @@
+package policyset
+
+const (
+	policySetResourceName   = "cyral_policy_set"
+	policySetDataSourceName = policySetResourceName
+)
