+++ cyral/internal/policy/wizard/constants.go
@@ -1,0 +1,5 @@
+package wizard
+
+const (
+	policyWizardsDataSourceName = "cyral_policy_wizards"
+)
