+++ cyral/internal/policy/wizard/schema_loader.go
@@ -1,4 +1,4 @@
-package policyset
+package wizard
 
 import "github.com/cyralinc/terraform-provider-cyral/cyral/core"
 
@@ -11,21 +11,10 @@ func (p *packageSchema) Name() string {
 
 func (p *packageSchema) Schemas() []*core.SchemaDescriptor {
 	return []*core.SchemaDescriptor{
-
-		{
-			Name:   policySetDataSourceName,
-			Type:   core.DataSourceSchemaType,
-			Schema: policySetDataSourceSchema,
-		},
 		{
 			Name:   policyWizardsDataSourceName,
 			Type:   core.DataSourceSchemaType,
 			Schema: policyWizardsDataSourceSchema,
-		},
-		{
-			Name:   policySetResourceName,
-			Type:   core.ResourceSchemaType,
-			Schema: policySetResourceSchema,
 		},
 	}
 }
