--- cyral/internal/policyset/constants.go
@@ -1,7 +1,0 @@
-package policyset
-
-const (
-	policySetResourceName       = "cyral_policy_set"
-	policySetDataSourceName     = policySetResourceName
-	policyWizardsDataSourceName = "cyral_policy_wizards"
-)
